backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-core
supabase
python-docx
pymupdf
openai
python-dotenv
pytest
httpx
faster-whisper
